Dynamic/tweetiment-scraping/requirements.txt
Click>=7.0
dash>=1.1.1
dash-core-components>=1.1.1
dash-html-components>=1.0.0
dash-renderer>=1.0.0
dash-table>=4.1.0
Flask>=1.1.1
Flask-Compress>=1.4.0
gunicorn>=19.9.0
itsdangerous>=1.1.0
Jinja2>=2.10.1
MarkupSafe>=1.1.1
plotly>=4.1.0
PyYAML>=5.1.2
retrying>=1.3.3
six>=1.12.0
Werkzeug>=0.15.5
pandas>=0.25.0
nltk>=3.4.4
textblob>=0.15.2
tweepy>=3.8.0
psycopg2-binary>=2.8.3
mysql-connector-python
mysql